refactor(dashboard): shared dashboard year window + SQL extract
Move last-five-years window and year/month extract fragments to
app/dashboard/_shared/dashboard-year-window.ts as dashboardYearWindow
and yearMonthSqlFromColumn. Payroll settled aggregates, advance repayment
aggregates, and timesheet hours aggregates import the shared helpers;
monthly-dashboard-aggregation keeps only settledPayrollOverlappingYearsWhere.
